Restructured gql folder
