+++ components/CartBox.tsx
@@ -1,6 +1,6 @@
 
 import { useMutation } from '@apollo/react-hooks'
-import CHANGE_IN_STOCK_ITEMS_IN_CART_MUTATION from '../constants/graphql/changeInStockItemsInCart.mutation'
+import CHANGE_IN_STOCK_ITEMS_IN_CART_MUTATION from '../graphql/mutations/changeInStockItemsInCart.mutation'
 
 import PrimaryButton from '../components/PrimaryButton'
