+++ graphql/queries/board.query.ts
