+++ graphql/queries/stains.query.ts
