+++ pages/boards/[id].tsx
@@ -3,8 +3,8 @@ import { useRouter } from 'next/router'
 import { useQuery, useMutation } from '@apollo/react-hooks'
 import { useState } from 'react'
 
-import CHANGE_IN_STOCK_ITEMS_IN_CART_MUTATION from '../../constants/graphql/changeInStockItemsInCart.mutation'
-import BOARD_QUERY from '../../constants/graphql/board.query'
+import CHANGE_IN_STOCK_ITEMS_IN_CART_MUTATION from '../../graphql/mutations/changeInStockItemsInCart.mutation'
+import BOARD_QUERY from '../../graphql/queries/board.query'
 
 import Layout from '../../components/Layout'
 import ReviewBox from '../../components/ReviewBox'
